+++ packages/eslint-plugin/src/rules/key-spacing.ts
@@ -1,5 +1,6 @@
 import type { TSESTree } from '@typescript-eslint/utils';
 import { AST_NODE_TYPES } from '@typescript-eslint/utils';
+import { getSourceCode } from '@typescript-eslint/utils/eslint-utils';
 
 import type {
   InferMessageIdsTypeFromRule,
@@ -23,16 +24,6 @@ const baseSchema = Array.isArray(baseRule.meta.schema)
   ? baseRule.meta.schema[0]
   : baseRule.meta.schema;
 
-/**
- * TODO: replace with native .at() once Node 14 stops being supported
- */
-function at<T>(arr: T[], position: number): T | undefined {
-  if (position < 0) {
-    return arr[arr.length + position];
-  }
-  return arr[position];
-}
-
 export default createRule<Options, MessageIds>({
   name: 'key-spacing',
   meta: {
@@ -50,7 +41,7 @@ export default createRule<Options, MessageIds>({
   defaultOptions: [{}],
 
   create(context, [options]) {
-    const sourceCode = context.getSourceCode();
+    const sourceCode = getSourceCode(context);
     const baseRules = baseRule.create(context);
 
     /**
@@ -59,7 +50,7 @@ export default createRule<Options, MessageIds>({
     function adjustedColumn(position: TSESTree.Position): number {
       const line = position.line - 1; // position.line is 1-indexed
       return getStringLength(
-        at(sourceCode.lines, line)!.slice(0, position.column),
+        sourceCode.lines.at(line)!.slice(0, position.column),
       );
     }
 
@@ -114,7 +105,7 @@ export default createRule<Options, MessageIds>({
       return code.slice(
         0,
         sourceCode.getTokenAfter(
-          at(node.parameters, -1)!,
+          node.parameters.at(-1)!,
           isClosingBracketToken,
         )!.range[1] - node.range[0],
       );
@@ -129,7 +120,7 @@ export default createRule<Options, MessageIds>({
       return getLastTokenBeforeColon(
         node.type !== AST_NODE_TYPES.TSIndexSignature
           ? node.key
-          : at(node.parameters, -1)!,
+          : node.parameters.at(-1)!,
       ).loc.end;
     }
 
@@ -230,7 +221,7 @@ export default createRule<Options, MessageIds>({
       if (
         leadingComments.length &&
         leadingComments[0].loc.start.line - groupEndLine <= 1 &&
-        candidateValueStartLine - at(leadingComments, -1)!.loc.end.line <= 1
+        candidateValueStartLine - leadingComments.at(-1)!.loc.end.line <= 1
       ) {
         for (let i = 1; i < leadingComments.length; i++) {
           if (
@@ -400,7 +391,7 @@ export default createRule<Options, MessageIds>({
         let prevNode: TSESTree.Node | undefined = undefined;
 
         for (const node of members) {
-          let prevAlignedNode = at(currentAlignGroup, -1);
+          let prevAlignedNode = currentAlignGroup.at(-1);
           if (prevAlignedNode !== prevNode) {
             prevAlignedNode = undefined;
           }