+++ packages/eslint-plugin/src/rules/naming-convention-utils/validator.ts
@@ -167,7 +167,7 @@ function createValidator(
       count,
       affixes: affixes?.join(', '),
       formats: formats?.map(f => PredefinedFormats[f]).join(', '),
-      regex: custom?.regex?.toString(),
+      regex: custom?.regex.toString(),
       regexMatch:
         custom?.match === true
           ? 'match'