+++ packages/eslint-plugin/src/rules/no-invalid-void-type.ts
@@ -1,5 +1,6 @@
 import type { TSESTree } from '@typescript-eslint/utils';
 import { AST_NODE_TYPES } from '@typescript-eslint/utils';
+import { getSourceCode } from '@typescript-eslint/utils/eslint-utils';
 
 import { createRule } from '../util';
 
@@ -93,15 +94,15 @@ export default createRule<[Options], MessageIds>({
       // extra check for precaution
       /* istanbul ignore next */
       if (
-        node.parent?.type !== AST_NODE_TYPES.TSTypeParameterInstantiation ||
-        node.parent.parent?.type !== AST_NODE_TYPES.TSTypeReference
+        node.parent.type !== AST_NODE_TYPES.TSTypeParameterInstantiation ||
+        node.parent.parent.type !== AST_NODE_TYPES.TSTypeReference
       ) {
         return;
       }
 
       // check whitelist
       if (Array.isArray(allowInGenericTypeArguments)) {
-        const sourceCode = context.getSourceCode();
+        const sourceCode = getSourceCode(context);
         const fullyQualifiedName = sourceCode
           .getText(node.parent.parent.typeName)
           .replace(/ /gu, '');
@@ -158,7 +159,7 @@ export default createRule<[Options], MessageIds>({
           (member.type === AST_NODE_TYPES.TSTypeReference &&
             member.typeArguments?.type ===
               AST_NODE_TYPES.TSTypeParameterInstantiation &&
-            member.typeArguments?.params
+            member.typeArguments.params
               .map(param => param.type)
               .includes(AST_NODE_TYPES.TSVoidKeyword)),
       );