+++ packages/eslint-plugin/src/rules/padding-line-between-statements.ts
@@ -1,5 +1,6 @@
 import type { TSESLint, TSESTree } from '@typescript-eslint/utils';
 import { AST_NODE_TYPES } from '@typescript-eslint/utils';
+import { getSourceCode } from '@typescript-eslint/utils/eslint-utils';
 
 import {
   createRule,
@@ -127,9 +128,7 @@ function newNodeTypeTester(type: AST_NODE_TYPES): NodeTestObject {
  * @private
  */
 function skipChainExpression(node: TSESTree.Node): TSESTree.Node {
-  return node && node.type === AST_NODE_TYPES.ChainExpression
-    ? node.expression
-    : node;
+  return node.type === AST_NODE_TYPES.ChainExpression ? node.expression : node;
 }
 
 /**
@@ -163,9 +162,9 @@ function isIIFEStatement(node: TSESTree.Node): boolean {
  */
 function isCJSRequire(node: TSESTree.Node): boolean {
   if (node.type === AST_NODE_TYPES.VariableDeclaration) {
-    const declaration = node.declarations[0];
+    const declaration = node.declarations.at(0);
     if (declaration?.init) {
-      let call = declaration?.init;
+      let call = declaration.init;
       while (call.type === AST_NODE_TYPES.MemberExpression) {
         call = call.object;
       }
@@ -234,8 +233,8 @@ function isDirective(
 ): boolean {
   return (
     node.type === AST_NODE_TYPES.ExpressionStatement &&
-    (node.parent?.type === AST_NODE_TYPES.Program ||
-      (node.parent?.type === AST_NODE_TYPES.BlockStatement &&
+    (node.parent.type === AST_NODE_TYPES.Program ||
+      (node.parent.type === AST_NODE_TYPES.BlockStatement &&
         isFunction(node.parent.parent))) &&
     node.expression.type === AST_NODE_TYPES.Literal &&
     typeof node.expression.value === 'string' &&
@@ -408,8 +407,7 @@ function verifyForNever(
       const nextToken = paddingLines[0][1];
       const start = prevToken.range[1];
       const end = nextToken.range[0];
-      const text = context
-        .getSourceCode()
+      const text = getSourceCode(context)
         .text.slice(start, end)
         .replace(PADDING_LINE_SEQUENCE, replacerToRemovePaddingLines);
 
@@ -445,7 +443,7 @@ function verifyForAlways(
     node: nextNode,
     messageId: 'expectedBlankLine',
     fix(fixer) {
-      const sourceCode = context.getSourceCode();
+      const sourceCode = getSourceCode(context);
       let prevToken = getActualLastToken(prevNode, sourceCode)!;
       const nextToken =
         sourceCode.getFirstTokenBetween(prevToken, nextNode, {
@@ -478,7 +476,7 @@ function verifyForAlways(
             }
             return true;
           },
-        })! || nextNode;
+        }) || nextNode;
       const insertText = isTokenOnSameLine(prevToken, nextToken)
         ? '\n\n'
         : '\n';
@@ -645,9 +643,9 @@ export default createRule<Options, MessageIds>({
   },
   defaultOptions: [],
   create(context) {
-    const sourceCode = context.getSourceCode();
+    const sourceCode = getSourceCode(context);
     // eslint-disable-next-line no-restricted-syntax -- We need all raw options.
-    const configureList = context.options || [];
+    const configureList = context.options;
 
     type Scope = {
       upper: Scope;