+++ packages/eslint-plugin/src/rules/prefer-enum-initializers.ts
@@ -1,4 +1,5 @@
 import type { TSESLint, TSESTree } from '@typescript-eslint/utils';
+import { getSourceCode } from '@typescript-eslint/utils/eslint-utils';
 
 import { createRule } from '../util';
 
@@ -23,7 +24,7 @@ export default createRule<[], MessageIds>({
   },
   defaultOptions: [],
   create(context) {
-    const sourceCode = context.getSourceCode();
+    const sourceCode = getSourceCode(context);
 
     function TSEnumDeclaration(node: TSESTree.TSEnumDeclaration): void {
       const { members } = node;