+++ packages/eslint-plugin/src/rules/return-await.ts
@@ -1,5 +1,6 @@
 import type { TSESLint, TSESTree } from '@typescript-eslint/utils';
 import { AST_NODE_TYPES } from '@typescript-eslint/utils';
+import { getSourceCode } from '@typescript-eslint/utils/eslint-utils';
 import * as tsutils from 'ts-api-utils';
 import * as ts from 'typescript';
 
@@ -54,7 +55,7 @@ export default createRule({
   create(context, [option]) {
     const services = getParserServices(context);
     const checker = services.program.getTypeChecker();
-    const sourceCode = context.getSourceCode();
+    const sourceCode = getSourceCode(context);
 
     const scopeInfoStack: ScopeInfo[] = [];
 
@@ -310,7 +311,7 @@ export default createRule({
         }
       },
       ReturnStatement(node): void {
-        const scopeInfo = scopeInfoStack[scopeInfoStack.length - 1];
+        const scopeInfo = scopeInfoStack.at(-1);
         if (!scopeInfo?.hasAsync || !node.argument) {
           return;
         }