+++ packages/typescript-estree/src/create-program/createProjectService.ts
@@ -23,6 +23,7 @@ export function createProjectService(
   tsconfigRootDir: string,
 ): ProjectServiceSettings {
   // We import this lazily to avoid its cost for users who don't use the service
+  // TODO: Once we drop support for TS<5.3 we can import from "typescript" directly
   const tsserver = require('typescript/lib/tsserverlibrary') as typeof ts;
 
   // TODO: see getWatchProgramsForProjects